Add setting to allow synchronous remote calls
Currently a new thread is created to make the remote calls, but if you already track the events using a background job you don't need to create new threads and it's possible that the worker process is 'done' and thus getting killed off before the remote call is made.

This commit provides a 'synchronous' setting to configure that. To turn synchronous calls on:

    Totango::Config[:synchronous] = true